refactored first page visit check into custom hook
