+++ apps/app/src/pages/ShoppingPage/ShoppingPage.tsx
@@ -1,32 +1,14 @@
-import { useState, useEffect, useRef } from 'react';
 import styles from './ShoppingPage.module.scss';
+import { ShoppingProvider } from '@/context/ShoppingContext';
+import { useIsFirstPageVisit } from '@/hooks/useIsFirstPageVisit';
 
 import ShoppingItems from './sections/ShoppingItems';
 import ShoppingHeader from './sections/ShoppingHeader';
 import ShoppingWelcome from './sections/ShoppingWelcome';
-import { ShoppingProvider } from '@/context/ShoppingContext';
 
 export const ShoppingPage = () => {
-  const [firstShopVisit, setFirstShopVisit] = useState<boolean>(false);
-  const isMounted = useRef<boolean>(false);
-
-  useEffect(() => {
-    if (isMounted.current) return;
-
-    isMounted.current = true;
-    const storedValue = JSON.parse(
-      localStorage.getItem('firstShopVisit') || 'null',
-    );
-
-    if (storedValue === null || storedValue === true) {
-      localStorage.setItem('firstShopVisit', 'true');
-      setFirstShopVisit(true);
-      return;
-    }
-
-    localStorage.setItem('firstShopVisit', 'false');
-    setFirstShopVisit(false);
-  }, []);
+  const { firstShopVisit, setFirstShopVisit } =
+    useIsFirstPageVisit('firstShopVisit');
 
   return (
     <div className={styles.wrapper}>
